chore(architecture): 将架构边界纳入可执行静态契约
