+++ package.json
@@ -14,7 +14,8 @@
     "type-check:tests": "tsc -p tsconfig.tests.json --noEmit",
     "type-check:scripts": "tsc -p tsconfig.scripts.json --noEmit",
     "type-check": "pnpm type-check:app && pnpm type-check:tests && pnpm type-check:scripts",
-    "check": "pnpm type-check && pnpm lint && pnpm db:check && pnpm test",
+    "architecture:check": "tsx scripts/architecture/check.ts",
+    "check": "pnpm type-check && pnpm lint && pnpm architecture:check && pnpm db:check && pnpm test",
     "verify": "pnpm check && env -u DATABASE_URL -u RIVALHUB_LOCAL_DATABASE_URL -u NEXT_PUBLIC_SUPABASE_URL -u NEXT_PUBLIC_SUPABASE_ANON_KEY -u SUPABASE_SERVICE_ROLE_KEY -u RIVALHUB_DB_TARGET pnpm build",
     "verify:services": "tsx scripts/db/local.ts verify-services",
     "db:generate": "drizzle-kit generate",
