+++ src/actions/matches/index.ts
@@ -1,3 +1,2 @@
 export { generateSchedule, initializeStage, createMatch } from "./schedule";
 export { recordMapResult, updateMatchStatus, updateMatchScheduledAt, updateMatchCompletionDeadline, batchSetCompletionDeadline, deleteMatch, correctMapScore, updateMatchCompletedAt, forfeitMatch, syncBracketMatches } from "./results";
-export { runMatchTimeAutoAwardCron } from "./scheduling";
