+++ src/components/admin/CompetitivePlatformCatalog.tsx
@@ -16,7 +16,7 @@ import {
   updateCompetitivePlatformRankLabel,
   updateCompetitivePlatformSeason,
 } from "@/actions/competitive-platform";
-import { resolveCatalogSeasonRoles, type CompetitivePlatformCatalogEntry } from "@/lib/competitive/catalog";
+import { resolveCatalogSeasonRoles, type CompetitivePlatformCatalogEntry } from "@/lib/competitive/catalog-contract";
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogBody, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
