+++ src/lib/discipline/contract.ts
@@ -1,0 +1,10 @@
+export type SanctionEffect =
+  | "registration_block"
+  | "roster_block"
+  | "match_participation_block";
+
+export const SANCTION_EFFECTS: readonly SanctionEffect[] = [
+  "registration_block",
+  "roster_block",
+  "match_participation_block",
+];
